api update job

+++ server/controllers/article.controller.js
@@ -72,32 +72,43 @@ router.get("/", async (req, res, next) => {
 router.delete('/delete/:articleId', async (req, res, next) => {
     const articleId = req.params.articleId;
 
-    // delete object in article comment
-    await db.ArticleComment.destroy({
-        where: {
-            ArticleId: articleId 
-        }
-    });
+    if (articleId) {
+        // delete object in article comment
+        await db.ArticleComment.destroy({
+            where: {
+                ArticleId: articleId 
+            }
+        })
+        .catch(err => {
+            return res.status(404).send({message: err.message});
+        })
 
-    // delete object in article reaction
-    await db.ArticleReaction.destroy({
-        where: {
-            articleId: articleId
-        }
-    });
+        // delete object in article reaction
+        await db.ArticleReaction.destroy({
+            where: {
+                articleId: articleId
+            }
+        })
+        .catch(err => {
+            return res.status(404).send({message: err.message});
+        })
 
-    // delete object in article
-    await db.Article.destroy({
-        where: {
-            id: articleId
-        }
-    })
-    .then(article => {
-        return res.send({message: "delete successfully"});
-    })
-    .catch(err => {
-        return res.status(500).send({message: err.message});
-    })
+        // delete object in article
+        await db.Article.destroy({
+            where: {
+                id: articleId
+            }
+        })
+        .then(article => {
+            return res.send({message: "delete successfully"});
+        })
+        .catch(err => {
+            return res.status(500).send({message: err.message});
+        })
+    }
+    else {
+        return res.status(404).send({message: "Invalidate value"});
+    }
 })
 
 module.exports = router
